refactor: change structure of the solution
