+++ Core/Core.csproj
@@ -1,0 +1,12 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp2.2</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\GoodNews.DAL\GoodNews.DAL.csproj" />
+    <ProjectReference Include="..\GoodNews.Data\GoodNews.Data.csproj" />
+  </ItemGroup>
+
+</Project>
