+++ Core/IUnitOfWork.cs
@@ -1,10 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using GoodNews.DAL.Entities;
 using GoodNews.DAL.Repository;
+using GoodNews.Data.Entities;
 
-namespace GoodNews.DAL.UnitOfWork
+namespace Core
 {
     public interface IUnitOfWork : IDisposable
     {
