--- GoodNews.BL/GoodNews.BL.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29324.140
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GoodNews.BL", "GoodNews.BL.csproj", "{250B8297-84AD-4E10-A8F9-CADC4E30BD7C}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GoodNews.DAL", "..\GoodNews.Data\GoodNews.DAL.csproj", "{F5A8156A-9774-462E-AFD2-101474D9D531}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{250B8297-84AD-4E10-A8F9-CADC4E30BD7C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{250B8297-84AD-4E10-A8F9-CADC4E30BD7C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{250B8297-84AD-4E10-A8F9-CADC4E30BD7C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{250B8297-84AD-4E10-A8F9-CADC4E30BD7C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F5A8156A-9774-462E-AFD2-101474D9D531}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F5A8156A-9774-462E-AFD2-101474D9D531}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F5A8156A-9774-462E-AFD2-101474D9D531}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F5A8156A-9774-462E-AFD2-101474D9D531}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {A8658C30-EAF7-4497-AC7A-1DDD0DAA1CF7}
-	EndGlobalSection
-EndGlobal
