+++ GoodNews.BL/Startup.cs
@@ -1,19 +1,15 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using Core;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.HttpsPolicy;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using GoodNews.DAL;
-using GoodNews.DAL.Entities;
-using GoodNews.DAL.UnitOfWork;
 using GoodNews.DAL.Repository;
+using GoodNews.Data.Entities;
+using Services;
 
 namespace GoodNews.BL
 {
