+++ GoodNews.DAL/Repository/EntityFrameworkRepository.cs
@@ -1,9 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
-using GoodNews.DAL.Entities;
+using GoodNews.Data.Entities;
 using Microsoft.EntityFrameworkCore;
 
 namespace GoodNews.DAL.Repository
