+++ GoodNews.Data/Entities/Entity.cs
@@ -1,9 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using Microsoft.EntityFrameworkCore.Metadata.Internal;
 
-namespace GoodNews.DAL.Entities
+namespace GoodNews.Data.Entities
 {
     public abstract class Entity
     {
